feat: replace (cloud) deploy + start instance modals with overlays
Closes #2552
