--- client/src/plugins/zeebe-plugin/deployment-plugin/DeploymentPluginModal.less
@@ -1,14 +1,0 @@
-:local(.DeploymentPluginModal) {
-  .icon {
-    vertical-align: text-top;
-  }
-
-  fieldset:first-child {
-    margin-top: 0px;
-  }
-
-  fieldset:last-child {
-    margin-bottom: 0px;
-    padding-bottom: 16px;
-  }
-}
