+++ client/src/plugins/zeebe-plugin/deployment-plugin/DeploymentPluginOverlay.less
@@ -1,0 +1,23 @@
+:local(.DeploymentPluginOverlay) {
+
+  .target_radio {
+    .form-check-inline {
+      display: grid;
+      grid-gap: 5px;
+      grid-auto-columns: auto;
+      grid-auto-flow: row;
+    }
+  }
+
+  input[type="radio"] + label {
+    margin: 0;
+  }
+
+  .toggle-switch {
+    margin-top: 10px;
+  }
+
+  .toggle-switch__label {
+    font-size: 13px !important;
+  }
+}
