+++ client/src/plugins/zeebe-plugin/start-instance-plugin/__tests__/StartInstancePluginSpec.js
@@ -38,8 +38,6 @@ describe('<StartInstancePlugin> (Zeebe)', () => {
       }
     });
 
-    console.log(wrapper.debug());
-
     // then
     expect(wrapper.find(BUTTON_SELECTOR)).to.have.lengthOf(1);
   });
